Added GLOBAL DELETE function
This function deletes a global variable that you own, permanently erasing it from the database (until it's created again) and freeing up the name to be used for someone else. This function exists because, the more global variables there are, the more global variables will start to slow down. Old B++ global variables were faster, so this function was less necessary there, but this is a problem in new B++, hence the addition of this function.